+++ src/cpu/ref_inner_product.cpp
@@ -186,13 +186,13 @@ void ref_inner_product_bwd_weights_t<data_type>::execute_backward_weights() {
     if (diff_bias) {
         diff_bias += diff_bias_d.blocking_desc().offset_padding;
         constexpr int blksize = 8;
-        size_t OC_blocks = OC / blksize;
+        int OC_blocks = OC / blksize;
         int rem_OC = OC % blksize;
 #       pragma omp parallel
         {
             const int ithr = omp_get_thread_num();
             const int nthr = omp_get_num_threads();
-            size_t oc_st{0}, oc_e{0};
+            int oc_st{0}, oc_e{0};
             balance211(OC_blocks, nthr, ithr, oc_st, oc_e);
             oc_st = oc_st * blksize;
             oc_e = oc_e * blksize;
