+++ src/cpu/simple_reorder.hpp
@@ -1229,10 +1229,11 @@ struct simple_reorder_impl<SIMPLE_REORDER_TEMPL_CALL,
             }
 
             if (rem_elems != 0 && ithr == nthr - 1){
-                for (int e = nelems - rem_elems; e < nelems; ++e){
-                    output[e] = data_t<type_o>((alpha == 1.0 && beta == 0.0) ?
-                            input[e]
-                            : alpha * input[e] + (beta ? beta * output[e] : 0));
+                for (size_t e = nelems - rem_elems; e < nelems; ++e) {
+                    output[e] = data_t<type_o>((alpha == 1.0 && beta == 0.0)
+                            ? input[e]
+                            : (alpha * input[e]
+                                + (beta ? beta * output[e] : 0)));
                 }
             }
         }
